% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/approx2.R
\name{approx2}
\alias{approx2}
\alias{approx3}
\alias{approx_InvQuantile}
\title{String of current time}
\usage{
approx2(x, y, xout, type)

approx3(x, y, xout)

approx_InvQuantile(x, y, yout, xmin, xmax)
}
\description{
These functions gives the current time
}
\examples{

}
\keyword{time}
