% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/divide_to_wide.R
\name{divide_to_wide}
\alias{divide_to_wide}
\title{String of current time}
\usage{
divide_to_wide(data, measure_var, signal_var, type_var, seed = 123)
}
\description{
These functions gives the current time
}
\examples{

}
\keyword{time}
