% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/is_finite_s.R
\name{is_finite_s}
\alias{is_finite_s}
\alias{is_na_s}
\title{Statistics of NAs}
\usage{
is_finite_s(x)

is_na_s(x)
}
\arguments{
\item{x}{is a numerical vector}
}
\description{
These functions calculate statistics of the number of 
finite and NAs in a given numerical vector
}
\examples{

}
\keyword{data.frame}
