% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/norm_vec_1.R
\name{norm_vec_1}
\alias{norm_vec_1}
\alias{norm_vec_2}
\title{String of current time}
\usage{
norm_vec_1(x)

norm_vec_2(x)
}
\description{
These functions gives the current time
}
\examples{

}
\keyword{time}
