% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rates2ODE.R
\name{rates2ODE}
\alias{rates2ODE}
\alias{ratestxt_process}
\alias{parse_stimulus}
\title{String of current time}
\usage{
rates2ODE(stoich, rates)

ratestxt_process(ratestxt)

parse_stimulus(rates, input_folder)
}
\description{
These functions gives the current time
}
\examples{

}
\keyword{time}
