% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{simulateSSA}
\alias{simulateSSA}
\alias{plot_traj}
\title{String of current time}
\usage{
simulateSSA(par, nu, x0, a, f_btl, tf)

plot_traj(data, tf, t_dens, path)
}
\description{
These functions gives the current time
}
\examples{

}
\keyword{time}
