% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary_extended.R
\name{summary_extended}
\alias{summary_extended}
\alias{summary_type2}
\alias{summary_type2_1}
\alias{summary_type3}
\title{Summaries functions}
\usage{
summary_extended(x)

summary_type2(x)

summary_type2_1(x)

summary_type3(x)
}
\arguments{
\item{x}{is a numeric vector whose statistics are to be calculated}
}
\value{
The summary of \code{x}
}
\description{
These functions provide different summary statistics of data
}
\examples{
summary_extended(x=rnorm(1000))
}
\keyword{summary}
