% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/t_test_meta.R
\name{t_test_meta}
\alias{t_test_meta}
\title{t-test for meta-analysis}
\usage{
t_test_meta(m1, m2, s1, s2, n1, n2)
}
\arguments{
\item{m1}{is mean of 1st group}

\item{m2}{is mean of 2nd group}

\item{s1}{is sd of 1st group}

\item{s2}{is sd of 2nd group}

\item{n1}{is size of 1st group}

\item{n2}{is size of 2nd group}
}
\value{
The p-value of t-test
}
\description{
This function conduct a t-test from mean-sd data
}
\examples{
t_test_meta(10,11,4,5,20,23)
}
\keyword{t-test}
