% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TJfunctions.R
\docType{package}
\name{TJfunctions}
\alias{TJfunctions}
\alias{TJfunctions-package}
\title{TJfunctions: a package with private utility functions}
\description{
Provide the basic easy-of-use functions for everyday use
}
\section{Utlity functions}{

tjfunctions
}

