% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trapz_simple.R
\name{trapz_simple}
\alias{trapz_simple}
\title{Numerical intergration by trapezoidal rule}
\usage{
trapz_simple(x, y)
}
\arguments{
\item{x}{is a numerical vector}

\item{y}{is a numerical vector}
}
\description{
This function calculate the integral of function
}
\examples{
trapz_simple(x=1:10,y=exp(1:10))
}
\keyword{integration}
\keyword{numerical}
